Safari-611.1.10/LayoutTests/editing/pasteboard/set_data_typeof_return.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test-pre.js"></script>
<script>
function setDataTest(e) {
    shouldBeUndefined("event.clipboardData.setData('text/plain', 'copy test')");
}
</script>
</head>
<body oncopy="setDataTest(event)">
<script>
description("Simple test to validate the return type of setData(). To run the test manually, simply select any text and initiate a copy operation.");

if (window.testRunner)
    window.testRunner.execCommand('copy');
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
